% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_env}
\alias{show_env}
\title{show_env}
\usage{
show_env()
}
\description{
show_env returns a list that contains two environments: calling, which is the
calling environment when you run show_env, and execution, which is the
execution environment when you run show_env.
}
