% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_envs}
\alias{show_envs}
\title{show_envs}
\usage{
show_envs()
}
\description{
show_envs returns a list that contains three environments: calling, which is
the calling environment when you run show_env; execution, which is the
execution environment when you run show_env; and enclosing, which is the
enclosing environment of show_envs.
}
