% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_paths}
\alias{show_paths}
\title{show_paths}
\usage{
show_paths()
}
\description{
show_paths returns a list of two environment lists for comparison.
}
\details{
call_stack contains an environment list that contains the labels of each
environment that is associated with the call stack that led to show_stack
being run. In other words, call_stack shows the execution environment of
show_paths as well as the execution environment of the function that called
show_paths, the function that called that function and so on.

search_path contains a list of every envrironment on the search path that
begins with the execution environment of show_paths. This is the path that R
would use to find objects called by show_path.
}
