% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_stack}
\alias{show_stack}
\title{show_stack}
\usage{
show_stack(e = parent.frame())
}
\arguments{
\item{e}{show_stack returns an environment list that contains the labels of each
environment that is associated with the call stack that led to show_stack
being run. In other words, show_stack returns the execution environment of
show_stack as well as the execution environment of the function that called
show_stack, the function that called that function and so on.}
}
\description{
show_stack
}
